Improved semantics
It is now forbidden to do things like length -= width. If implemented, the return type would be length&. However, length - width should give delta length. To avoid inconsistency, things like length -= width aren't implemented.

However, it does seem useful to supply += to do things like length += width. Note that operator+= is implementable in terms of operator+, which is encouraging.

This all said, there it is somewhat counterintuitive that += is allowed but -= isn't. However, at least for now, this seems to be the best design.